+++ CHANGELOG.md
@@ -15,6 +15,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 - Added `IntoIterator` implementations for `AlignmentStates` as well as an `iter()` method
 - Added `PartialEq` between `Cigar` and `AlignmentStates` without allocating on the heap
 - Added `score_from_path` for `GlobalPhmm`
+- Added `ref_len` and `query_len` to the `Alignment` struct
 
 ### Changed